Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/referrer-strict-policies.sub-expected.txt


FAIL The strict-* referrer policies compare the trustworthiness of a request's referrer string against its URL assert_equals: A document with the strict-origin referrer policy served over HTTP, imports an module script over HTTPS, that imports a descendant script over HTTP. The request for the descendant script is sent with no `Referer` header expected "" but got "http://localhost:8800/"
